feat: memberBoard의 수정과 삭제 기능 구현, localStorage, sessionStorage 수정/삭제 시 관리하기

+++ movie_project/src/components/MyPageBody.js
@@ -110,8 +110,14 @@ const MyPageBody = () => {
             .catch((e) => {
               console.error(e);
             });
+
+          if (
+            window.localStorage.getItem("id") ===
+            window.sessionStorage.getItem("id")
+          ) {
+            window.localStorage.clear();
+          }
           window.sessionStorage.clear();
-          window.localStorage.clear();
 
           navigate("/", { return: true });
         } else {
